LPS-3271 Delete and update asset only if group is a community
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@31152 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portal/service/impl/GroupLocalServiceImpl.java
@@ -397,7 +397,9 @@ public class GroupLocalServiceImpl extends GroupLocalServiceBaseImpl {
 
 		// Tags
 
-		tagsAssetLocalService.deleteAsset(Group.class.getName(), groupId);
+		if (group.isCommunity()) {
+			tagsAssetLocalService.deleteAsset(Group.class.getName(), groupId);
+		}
 
 		// Wiki
 
@@ -746,7 +748,7 @@ public class GroupLocalServiceImpl extends GroupLocalServiceBaseImpl {
 
 		// Tags
 
-		if (serviceContext != null) {
+		if ((serviceContext != null) && group.isCommunity()) {
 			updateTagsAsset(
 				group.getCreatorUserId(), group,
 				serviceContext.getTagsCategories(),
